+++ src/components/features/homepage/index.ts
@@ -1,8 +1,8 @@
-export { default as HeroSection } from './HeroSection'
-export { default as ProfessionSection } from './ProfessionSection'
-export { default as FeedbackSection } from './FeedbackSection'
-export { default as ActivitySection } from './ActivitySection'
-export { default as ImpactSection } from './ImpactSection'
-export { default as ContactSection } from './ContactSection'
-export { default as GallerySection } from './GallerySection'
-export { default as CTASection } from './CTASection'
+export { default as HeroSection } from '@/components/features/homepage/HeroSection'
+export { default as ProfessionSection } from '@/components/features/homepage/ProfessionSection'
+export { default as FeedbackSection } from '@/components/features/homepage/FeedbackSection'
+export { default as ActivitySection } from '@/components/features/homepage/ActivitySection'
+export { default as ImpactSection } from '@/components/features/homepage/ImpactSection'
+export { default as ContactSection } from '@/components/features/homepage/ContactSection'
+export { default as GallerySection } from '@/components/features/homepage/GallerySection'
+export { default as CTASection } from '@/components/features/homepage/CTASection'
