+++ src/components/ui/index.ts
@@ -1,15 +1,15 @@
-export { default as StatCard } from './StatCard'
-export { default as ProfessionGroup } from './ProfessionGroup'
-export { default as FeedbackCard } from './FeedbackCard'
-export * from './alert-dialog'
-export * from './button'
-export * from './calendar'
-export * from './card'
-export * from './dialog'
-export * from './form'
-export * from './input'
-export * from './label'
-export * from './popover'
-export * from './select'
-export * from './table'
-export * from './textarea'
+export { default as StatCard } from '@/components/ui/StatCard'
+export { default as ProfessionGroup } from '@/components/ui/ProfessionGroup'
+export { default as FeedbackCard } from '@/components/ui/FeedbackCard'
+export * from '@/components/ui/alert-dialog'
+export * from '@/components/ui/button'
+export * from '@/components/ui/calendar'
+export * from '@/components/ui/card'
+export * from '@/components/ui/dialog'
+export * from '@/components/ui/form'
+export * from '@/components/ui/input'
+export * from '@/components/ui/label'
+export * from '@/components/ui/popover'
+export * from '@/components/ui/select'
+export * from '@/components/ui/table'
+export * from '@/components/ui/textarea'
